# backend/docs/swagger.yaml
basePath: /
definitions:
  handlers.LoginRequest:
    properties:
      email:
        type: string
      password:
        type: string
    required:
    - email
    - password
    type: object
  handlers.RegisterRequest:
    properties:
      email:
        type: string
      name:
        type: string
      password:
        minLength: 6
        type: string
    required:
    - email
    - name
    - password
    type: object
  models.ClothingItem:
    properties:
      category:
        description: e.g., "Outerwear", "Top", "Bottom"
        type: string
      colors:
        description: Tags
        items:
          type: string
        type: array
      createdAt:
        description: Timestamps
        type: string
      description:
        description: |-
          The "Vibe" Engine
          Gemini will generate a text description, which we then convert to a vector
        type: string
      id:
        type: string
      imageUrl:
        description: Image Data
        type: string
      name:
        description: Basic Metadata
        type: string
      occasions:
        description: Casual, Formal
        items:
          type: string
        type: array
      seasons:
        description: Winter, Summer
        items:
          type: string
        type: array
      updatedAt:
        type: string
      userId:
        description: Good for scaling later
        type: string
    type: object
  router.PingResponse:
    properties:
      message:
        example: pong
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: API for Armoire application
  termsOfService: http://swagger.io/terms/
  title: Armoire API
  version: "1.0"
paths:
  /auth/login:
    post:
      consumes:
      - application/json
      description: Authenticate user and return JWT token
      parameters:
      - description: Login credentials
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Login successful with token and user data
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid request body
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Invalid email or password
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Log into an account
      tags:
      - auth
  /auth/register:
    post:
      consumes:
      - application/json
      description: Create a new user account with name, email and password
      parameters:
      - description: Registration details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.RegisterRequest'
      produces:
      - application/json
      responses:
        "201":
          description: User created successfully
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Invalid request body
          schema:
            additionalProperties:
              type: string
            type: object
        "409":
          description: Email already registered
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Register a new user
      tags:
      - auth
  /clothing/upload:
    post:
      consumes:
      - multipart/form-data
      description: Upload a clothing item image, analyze it with AI for auto-tagging,
        generate vector embeddings, and store in database
      parameters:
      - description: Clothing item image (max 10MB)
        in: formData
        name: image
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: Successfully uploaded and processed clothing item
          schema:
            $ref: '#/definitions/models.ClothingItem'
        "400":
          description: Invalid file
          schema:
            type: string
        "500":
          description: Failed to upload to GCS / AI Analysis Failed / Vector Embedding
            Failed / Database Save Failed
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Upload a clothing item
      tags:
      - clothing
  /ping:
    get:
      consumes:
      - application/json
      description: Returns pong message
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/router.PingResponse'
      summary: Ping endpoint
      tags:
      - health
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"